moving stars comment stuff to correct package
